Revert "refactor: 캘린더 컴포넌트 경로 수정"
This reverts commit 80311b62e490c2cfb5a5cafb280d3458b672fe5c.
